etl/requirements.txt
certifi==2022.9.24
elastic-transport==8.4.0
elasticsearch==8.4.3
psycopg2-binary==2.9
python-dateutil==2.8.2
python-dotenv==0.21.0
six==1.16.0
urllib3==1.26.12
